Fix-
Fixed issue where the member 'AbsoluteID' was being serialized to the CSV file.

Corrected a potential issues of manipulating the original Records collection during exporting. This was done by introducing a deep copy of Records and use this copy to exclude 'AbsoluteID' and seriliaze the data to the CSV file